requirement.txt
Django==3.2.13
djangorestframework==3.13.1
django-rest-swagger==2.2.0
requests>=2.27.1
django-filter-21.1